# Conflicts:
#	packages/kbn-test/src/functional_test_runner/lib/config/schema.ts
#	test/functional/page_objects/discover_page.js
#	x-pack/test/functional/page_objects/security_page.js

+++ x-pack/plugins/actions/server/routes/delete.ts
@@ -41,6 +41,9 @@ export const deleteActionRoute = (router: IRouter, licenseState: LicenseState) =
       res: KibanaResponseFactory
     ): Promise<IKibanaResponse<any>> {
       verifyApiAccess(licenseState);
+      if (!context.actions) {
+        return res.badRequest({ body: 'RouteHandlerContext is not registered for actions' });
+      }
       const actionsClient = context.actions.getActionsClient();
       const { id } = req.params;
       await actionsClient.delete({ id });
